+++ tests/testthat/test-reorder.R
@@ -33,14 +33,14 @@ test_that("can reorder by 2d summary", {
   expect_equal(levels(f2), c("a", "b"))
 })
 
-test_that("complains if summary doesn't return single number", {
+test_that("complains if summary doesn't return single value", {
   fun1 <- function(x, y) c(1, 2)
-  fun2 <- function(x, y) "a"
+  fun2 <- function(x, y) integer()
 
-  expect_error(fct_reorder("a", 1, fun1), "single number per group")
-  expect_error(fct_reorder("a", 1, fun2), "single number per group")
-  expect_error(fct_reorder2("a", 1, 2, fun1), "single number per group")
-  expect_error(fct_reorder2("a", 1, 2, fun2), "single number per group")
+  expect_error(fct_reorder("a", 1, fun1), "single value per group")
+  expect_error(fct_reorder("a", 1, fun2), "single value per group")
+  expect_error(fct_reorder2("a", 1, 2, fun1), "single value per group")
+  expect_error(fct_reorder2("a", 1, 2, fun2), "single value per group")
 })
 
 test_that("fct_infreq respects missing values", {